feat: begin work on CSS code explanations (read more)
1. Update snippet getStaticProps to include mapping of CSS to code
explanation
2. Update pre + code styles to work with divs. Prism expects
specifically a code block within pre. When we have a div as the
intermediary, that relationship is lost, and we lose stuff like
white-space preservation, because now Prism sees it as inline code (code
element on its own is seen as inline). Needed to manually add back
styles to compensate.
3. Began adding logic for expanding code explanations. Need to update
ResizableSplitColumns to ensure consistent widths.